// shared/api/command/CommandInt.ts
export interface CommandInt {
    readonly [x: string]: unknown;
    readonly iamName: string | undefined;
    readonly iamProfile: string | undefined;
    readonly acctName: string | undefined;
    readonly role: string | undefined;
    readonly region: string | undefined;
    readonly iamMfa: string | undefined;
    readonly _: string[];
    readonly $0: string;
}